requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
gunicorn==20.1.0
python-dateutil==2.8.2
google-cloud-storage==2.10.0